all: Move ApiSchema into graph::schema
