+++ graph/src/schema/mod.rs
@@ -1,5 +1,5 @@
 /// Generate full-fledged API schemas from existing GraphQL schemas.
-pub mod api;
+mod api;
 
 /// Utilities for working with GraphQL schema ASTs.
 pub mod ast;
@@ -9,5 +9,6 @@ mod input_schema;
 
 pub use api::{api_schema, APISchemaError};
 
+pub use api::{ApiSchema, ErrorPolicy};
 pub use fulltext::{FulltextAlgorithm, FulltextConfig, FulltextDefinition, FulltextLanguage};
 pub use input_schema::InputSchema;
